+++ src/app/component/arb/validation/arb_validation.service.ts
@@ -3,14 +3,14 @@ import * as he from "he";
 import path from "path";
 import * as vscode from "vscode";
 import { Cmd } from "../../../cmd/cmd";
-import { TextTranslateCmdArgs } from "../../../cmd/translation/text.translate.cmd";
 import { Dialog } from "../../../util/dialog";
-import { Editor } from "../../../util/editor";
 import { Link } from "../../../util/link";
 import { Toast } from "../../../util/toast";
-import { ConfigService } from "../../config/config";
+import { ConfigService, LanguageCode } from "../../config/config";
 import { Language } from "../../language/language";
 import { LanguageService } from "../../language/language.service";
+import { TranslationResult } from "../../translation/translation";
+import { TranslationService } from "../../translation/translation.service";
 import { ARB, ARBService } from "../arb";
 import { InvalidType, ValidationResult } from "./arb_validation";
 import { ARBValidationRepository } from "./arb_validation.repository";
@@ -19,23 +19,33 @@ interface InitParams {
   arbService: ARBService;
   configService: ConfigService;
   languageService: LanguageService;
+  translationService: TranslationService;
   arbValidationRepository: ARBValidationRepository;
+}
+
+enum RetranslateScope {
+  selection = "selection",
+  type = "type",
+  all = "all",
 }
 
 export class ARBValidationService {
   private arbService: ARBService;
   private configService: ConfigService;
   private languageService: LanguageService;
+  private translationService: TranslationService;
   private arbValidationRepository: ARBValidationRepository;
   constructor({
     arbService,
     configService,
     languageService,
+    translationService,
     arbValidationRepository,
   }: InitParams) {
     this.arbService = arbService;
     this.configService = configService;
     this.languageService = languageService;
+    this.translationService = translationService;
     this.arbValidationRepository = arbValidationRepository;
   }
 
@@ -60,7 +70,10 @@ export class ARBValidationService {
     return validationResults;
   }
 
-  public async validate(validationResult: ValidationResult): Promise<boolean> {
+  public async validate(
+    validationResult: ValidationResult,
+    validationResultList: ValidationResult[]
+  ): Promise<boolean> {
     const { sourceARB, targetARB, invalidType, invalidMessage, key } =
       validationResult;
     const targetFileName = path.basename(targetARB.filePath);
@@ -74,43 +87,67 @@ export class ARBValidationService {
       validationResult.key
     );
 
-    switch (validationResult.invalidType) {
-      case InvalidType.notExcluded:
-        // text translate
-        const yes = await Dialog.showConfirmDialog({
-          title: "Excluded Text Not Found",
-          placeHolder: "Do you want to translate again without cache?",
-        });
-        if (yes) {
-          const { editor: targetEditor } = await Editor.open(
-            targetARB.filePath,
-            vscode.ViewColumn.Two
-          );
-          const selection = Editor.selectFromARB(
-            targetEditor,
-            key,
-            `${targetARB.data[key]}`
-          );
-          await vscode.commands.executeCommand(Cmd.TextTranslate, <
-            TextTranslateCmdArgs
-          >{
-            queries: [sourceARB.data[key]],
-            selections: [selection],
-            sourceLang: sourceARB.language,
-            targetLang: targetARB.language,
-            useCache: false,
-          });
-        }
-        break;
+    switch (invalidType) {
       case InvalidType.keyNotFound:
+      case InvalidType.notExcluded:
+      case InvalidType.invalidLineBreaks:
       case InvalidType.invalidParameters:
       case InvalidType.invalidParentheses:
-        // open translation website
-        await Link.openGoogleTranslateWebsite({
-          sourceLanguage: sourceARB.language,
-          targetLanguage: validationResult.targetARB.language,
-          text: sourceARB.data[validationResult.key],
-        });
+        const action = await this.selectNextAction(invalidType);
+        if (!action) {
+          return false;
+        }
+        if (action === Cmd.GoogleTranslationOpenWeb) {
+          // open translation website
+          await Link.openGoogleTranslateWebsite({
+            sourceLanguage: sourceARB.language,
+            targetLanguage: validationResult.targetARB.language,
+            text: sourceARB.data[validationResult.key],
+            isConfirm: false,
+          });
+        } else {
+          // text translation
+          const sameTypeValidationResults = validationResultList.filter(
+            (v) => v.invalidType === validationResult.invalidType
+          );
+          const retranslateScope = await this.selectRetranslateScope({
+            nTotal: validationResultList.length,
+            nType: sameTypeValidationResults.length,
+            invalidType,
+          });
+          if (!retranslateScope) {
+            return false;
+          }
+
+          let translateResult: TranslationResult;
+          switch (retranslateScope) {
+            case RetranslateScope.selection:
+              // retranslate only selected item
+              translateResult = await this.translationService.translate({
+                queries: [sourceARB.data[key]],
+                sourceLang: sourceARB.language,
+                targetLang: targetARB.language,
+                useCache: false,
+                isEncodeARBParams: true,
+              });
+              targetARB.data[key] = translateResult.data[0];
+              this.arbService.upsert(targetARB.filePath, targetARB.data);
+              break;
+            case RetranslateScope.type:
+              // translate [validationResult.invalidType]
+              translateResult = await this.retranslateAll(
+                sameTypeValidationResults
+              );
+              break;
+            case RetranslateScope.all:
+              // retranslate all
+              translateResult = await this.retranslateAll(validationResultList);
+              break;
+          }
+          Toast.i(
+            `🟢 Re-translation Complete. (API: ${translateResult.nAPICall.toLocaleString()}, Cache: ${translateResult.nCache.toLocaleString()})`
+          );
+        }
         break;
       case InvalidType.undecodedHtmlEntityExists:
         // decode html entities
@@ -124,6 +161,136 @@ export class ARBValidationService {
         break;
     }
     return false;
+  }
+
+  private async selectNextAction(
+    invalidType: InvalidType
+  ): Promise<Cmd | undefined> {
+    return (
+      await vscode.window.showQuickPick(
+        [
+          {
+            label: "Re-translate without cache.",
+            data: Cmd.TextTranslate,
+          },
+          {
+            label: "Open Google Translate website",
+            data: Cmd.GoogleTranslationOpenWeb,
+          },
+        ],
+        {
+          ignoreFocusOut: true,
+          title: invalidType,
+          placeHolder: "Please select the next action.",
+        }
+      )
+    )?.data;
+  }
+
+  private async selectRetranslateScope({
+    nType,
+    nTotal,
+    invalidType,
+  }: {
+    nType: number;
+    nTotal: number;
+    invalidType: InvalidType;
+  }): Promise<RetranslateScope | undefined> {
+    return (
+      await vscode.window.showQuickPick(
+        [
+          {
+            label: "Selected item (1)",
+            data: RetranslateScope.selection,
+          },
+          {
+            label: `Selected types - ${invalidType} (${nType.toLocaleString()})`,
+            data: RetranslateScope.type,
+          },
+          {
+            label: `All (${nTotal.toLocaleString()})`,
+            data: RetranslateScope.all,
+          },
+        ],
+        {
+          ignoreFocusOut: true,
+          title: "Re-translation Scope",
+          placeHolder: `Select re-translate scope`,
+        }
+      )
+    )?.data;
+  }
+
+  private async retranslateAll(
+    validationResultList: ValidationResult[]
+  ): Promise<TranslationResult> {
+    let nComplete: number = 0;
+    const translateResult: TranslationResult = {
+      data: [],
+      nAPICall: 0,
+      nCache: 0,
+    };
+    await vscode.window.withProgress(
+      {
+        location: vscode.ProgressLocation.Notification,
+        cancellable: true,
+      },
+      async (progress, token) => {
+        // Group by same target language
+        const targetGroup: {
+          [targetLangCode: LanguageCode]: ValidationResult[];
+        } = {};
+        let nGroup = 0;
+        for (const v of validationResultList) {
+          const targetLangCode = v.targetARB.language.languageCode;
+          if (!targetGroup[targetLangCode]) {
+            targetGroup[targetLangCode] = [];
+            nGroup += 1;
+          }
+          targetGroup[targetLangCode].push(v);
+        }
+
+        for (const targetLangCode of Object.keys(targetGroup)) {
+          if (token.isCancellationRequested) {
+            // cancel
+            Toast.i(`🟠 Canceled`);
+            break;
+          }
+          const sameTarget = targetGroup[targetLangCode];
+          const { sourceARB, targetARB } = sameTarget[0];
+          const queries = sameTarget.map((v) => v.sourceARB.data[v.key]);
+
+          nComplete += queries.length;
+          progress.report({
+            increment: 100 / nGroup,
+            message: `${nComplete} / ${validationResultList.length} re-translating...`,
+          });
+
+          // Translation
+          const result = await this.translationService.translate({
+            queries,
+            sourceLang: sourceARB.language,
+            targetLang: targetARB.language,
+            useCache: false,
+            isEncodeARBParams: true,
+          });
+          translateResult.data.push(...result.data),
+            (translateResult.nCache += result.nCache);
+          translateResult.nAPICall += result.nAPICall;
+          const translatedTextList = result.data;
+
+          // Upsert
+          for (let i = 0; i < sameTarget.length; i++) {
+            const { key } = sameTarget[i];
+            const translatedText = translatedTextList[i];
+            targetARB.data[key] = translatedText;
+          }
+          this.arbService.upsert(targetARB.filePath, targetARB.data);
+        }
+      }
+    );
+
+    return translateResult;
   }
 
   public async decodeHtmlEntities(targetArb: ARB, keyList: string[]) {
@@ -144,7 +311,7 @@ export class ARBValidationService {
   ): AsyncGenerator<ValidationResult, undefined, ValidationResult> {
     // get source ParamsValidation
     const sourceValidation =
-      this.arbValidationRepository.getParamsValidation(sourceArb);
+      this.arbValidationRepository.getValidation(sourceArb);
     if (Object.keys(sourceValidation).length === 0) {
       return;
     }
@@ -166,7 +333,7 @@ export class ARBValidationService {
       // get targetArb
       const targetArb: ARB = await this.arbService.getARB(targetArbFilePath);
       const targetValidation =
-        this.arbValidationRepository.getParamsValidation(targetArb);
+        this.arbValidationRepository.getValidation(targetArb);
 
       // generate validation result
       yield* this.arbValidationRepository.generateValidationResult(