+++ src/app/component/arb/validation/arb_validation.ts
@@ -1,26 +1,21 @@
+import { TextStatistic } from "../../../util/statistic";
 import { ARB } from "../arb";
 
-export interface ARBValidationData {
-  value: string;
-  nParams: number;
-  nParentheses: number;
-  nHtmlEntities: number;
-}
-
 export interface ARBValidation {
-  [key: string]: ARBValidationData;
+  [key: string]: TextStatistic;
 }
 
 export enum InvalidType {
   notExcluded = "Not excluded",
   keyNotFound = "Key does not exist",
+  invalidLineBreaks = "Incorrect number of line breaks",
   invalidParameters = "Incorrect number of parameters",
   invalidParentheses = "Incorrect number of parentheses",
   undecodedHtmlEntityExists = "Undecoded html entity exists",
 }
 
 export interface ValidationResult {
-  sourceValidationData: ARBValidationData;
+  sourceValidationData: TextStatistic;
   invalidType: InvalidType;
   invalidMessage?: string;
   sourceARB: ARB;