+++ src/app/util/constant.ts
@@ -3,9 +3,20 @@ export class Constant {
   public static homePage =
     "https://marketplace.visualstudio.com/items?itemName=DevStory.flutter-translator";
 
-  public static paramReplaceKeys: string[] = [
+  public static keycaps: string[] = [
+    "0️⃣",
+    "1️⃣",
+    "2️⃣",
+    "3️⃣",
+    "4️⃣",
+    "5️⃣",
+    "6️⃣",
+    "7️⃣",
+    "8️⃣",
+    "9️⃣",
+  ];
+  public static emojis: string[] = [
     "😀",
-    "😃",
     "😄",
     "😁",
     "🥹",